-Werror -Wall and -Wextra

+++ source/system/memory/memory.cpp
@@ -6,7 +6,7 @@ uint64_t getmemsize()
     static uint64_t meminbytes = 0;
     if (meminbytes > 0) return meminbytes;
 
-    for (int i = 0; i < mmap_tag->entries; i++)
+    for (uint64_t i = 0; i < mmap_tag->entries; i++)
     {
         if (mmap_tag->memmap[i].type != STIVALE2_MMAP_USABLE) continue;
